# perf(txindex): search optimization