--- a/.changelog/unreleased/improvements/3458-txindex-search-optimization.md
+++ b/.changelog/unreleased/improvements/3458-txindex-search-optimization.md
@@ -1,0 +1,2 @@
+- `[state/txindex]` search optimization
+  ([\#3458](https://github.com/cometbft/cometbft/pull/3458))