--- a/state/txindex/kv/kv.go
+++ b/state/txindex/kv/kv.go
@@ -27,9 +27,10 @@ import (
 )
 
 const (
-	tagKeySeparator     = "/"
-	tagKeySeparatorRune = '/'
-	eventSeqSeparator   = "$es$"
+	tagKeySeparator          = "/"
+	tagKeySeparatorRune      = '/'
+	eventSeqSeparator        = "$es$"
+	eventSeqSeperatorRuneAt0 = '$'
 )
 
 var (
@@ -406,6 +407,38 @@ func (txi *TxIndex) indexEvents(result *abci.TxResult, hash []byte, store dbm.Ba
 	return nil
 }
 
+type hashKey struct {
+	hash   string
+	height int64
+}
+
+type hashKeySorter struct {
+	keys []hashKey
+	by   func(t1, t2 *hashKey) bool
+}
+
+func (t *hashKeySorter) Less(i, j int) bool { return t.by(&t.keys[i], &t.keys[j]) }
+func (t *hashKeySorter) Len() int           { return len(t.keys) }
+func (t *hashKeySorter) Swap(i, j int)      { t.keys[i], t.keys[j] = t.keys[j], t.keys[i] }
+
+func byHeightDesc(i, j *hashKey) bool {
+	hi := i.height
+	hj := j.height
+	if hi == hj {
+		return i.hash > j.hash
+	}
+	return hi > hj
+}
+
+func byHeightAsc(i, j *hashKey) bool {
+	hi := i.height
+	hj := j.height
+	if hi == hj {
+		return i.hash > j.hash
+	}
+	return hi > hj
+}
+
 // Search performs a search using the given query.
 //
 // It breaks the query into conditions (like "tx.height > 5"). For each
@@ -515,26 +548,23 @@ func (txi *TxIndex) Search(ctx context.Context, q *query.Query, pagSettings txin
 	numResults := len(filteredHashes)
 
 	// Convert map keys to slice for deterministic ordering
-	hashKeys := make([]string, 0, numResults)
-	for k := range filteredHashes {
-		hashKeys = append(hashKeys, k)
+	hashKeys := make([]hashKey, 0, numResults)
+	for k, v := range filteredHashes {
+		hashKeys = append(hashKeys, hashKey{hash: k, height: v.Height})
+	}
+
+	var by func(i, j *hashKey) bool
+
+	if pagSettings.OrderDesc {
+		by = byHeightDesc
+	} else {
+		by = byHeightAsc
 	}
 
 	// Sort by height
-	sort.Slice(hashKeys, func(i, j int) bool {
-		hi := filteredHashes[hashKeys[i]].Height
-		hj := filteredHashes[hashKeys[j]].Height
-		if hi == hj {
-			// If heights are equal, sort lexicographically
-			if pagSettings.OrderDesc {
-				return hashKeys[i] > hashKeys[j]
-			}
-			return hashKeys[i] < hashKeys[j]
-		}
-		if pagSettings.OrderDesc {
-			return hi > hj
-		}
-		return hi < hj
+	sort.Sort(&hashKeySorter{
+		keys: hashKeys,
+		by:   by,
 	})
 
 	// If paginated, determine which hash keys to return
@@ -565,7 +595,7 @@ func (txi *TxIndex) Search(ctx context.Context, q *query.Query, pagSettings txin
 	resultMap := make(map[string]struct{})
 RESULTS_LOOP:
 	for _, hKey := range hashKeys {
-		h := filteredHashes[hKey].TxBytes
+		h := filteredHashes[hKey.hash].TxBytes
 		res, err := txi.Get(h)
 		if err != nil {
 			return nil, 0, fmt.Errorf("failed to get Tx{%X}: %w", h, err)
@@ -995,13 +1025,21 @@ func extractValueFromKey(key []byte) string {
 }
 
 func extractEventSeqFromKey(key []byte) string {
-	parts := strings.Split(string(key), tagKeySeparator)
+	endPos := bytes.LastIndexByte(key, tagKeySeparatorRune)
 
-	lastEl := parts[len(parts)-1]
-
-	if strings.Contains(lastEl, eventSeqSeparator) {
-		return strings.SplitN(lastEl, eventSeqSeparator, 2)[1]
+	if endPos == -1 {
+		return "0"
 	}
+
+	for ; endPos < len(key); endPos++ {
+		if key[endPos] == eventSeqSeperatorRuneAt0 && len(key)-endPos > 4 {
+			eventSeq := string(key[endPos:])
+			if eventSeq, ok := strings.CutPrefix(eventSeq, eventSeqSeparator); ok {
+				return eventSeq
+			}
+		}
+	}
+
 	return "0"
 }