@melekes 2024-07-10
> Really helpful, thank you! Can we backport this?

According to our backporting policy, we will only backport bug fixes to v0.38 and v0.37. We can backport to the `v0.37.x-experimental` and/or `v0.38.x-experimental` branches if needed.